% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipegs.R
\docType{package}
\name{pipegs}
\alias{pipegs}
\alias{pipegs-package}
\title{pvalue approximation with stolarsky invariance principal.}
\description{
The package calculates the approximation for permutation p-values of linear statistic
}

